Reorganized everything.
mklink is now also a library crate,
and it has an API better for library use
with more strict, typed errors and intermediate structs.

Pre-link error checking is also done on the WSL side,
to avoid an exec() if necessary.
Since it's meant to be used in a library as well,
spawn() is used instead of exec().
Errors on the Windows side are also marshalled to the WSL in stdout using bincode.

The error messages aren't prettified yet (just debug prints).